Separated components to different files
