+++ src/App.jsx
@@ -1,25 +1,27 @@
 import React, { useState } from 'react'; // Import useState
 import birdData from './data/birds'; // Import bird data
 import bonusItems from './data/bonusItems';
+import Cards from "./components/Cards.jsx"
+import Cart from "./components/Cart.jsx"
 
 // Step 1: import useStates
 // Step 2: import Data
 // Step 3: make Cards function to get data and render the information on the page
 // Step 3a: put Cards fx as prop in App function in <main> tag
-function Cards ({ birdData, addToCartProp }) { // name of prop is addToCart
-  return (
-    <section>
-      {birdData.map((bird) => (
-        <div key={bird.id} className="card">
-          <h3>{bird.name}</h3>
-          <img src={bird.img} alt={bird.name} />
-          <p>${bird.amount.toFixed(2)}</p>
-          <button onClick={() => addToCartProp(bird)}>Adopt</button>
-        </div>
-      ))}
-    </section>
-  );
-}
+// function Cards ({ birdData, addToCartProp }) { // name of prop is addToCart
+//   return (
+//     <section className='birds'>
+//       {birdData.map((bird) => (
+//         <div key={bird.id} className="card">
+//           <h3>{bird.name}</h3>
+//           <img src={bird.img} alt={bird.name} />
+//           <p>${bird.amount.toFixed(2)}</p>
+//           <button onClick={() => addToCartProp(bird)}>Adopt</button>
+//         </div>
+//       ))}
+//     </section>
+//   );
+// }
 
 // Step 4: Make Cart component
 
@@ -29,52 +31,52 @@ Step 5a: create variable to store totalCost and use as a (totalCostProp)
 Step 5b: create variable to store discount rate to use as data for totalCostProp
 Step 5c: create variable to store discountedTotal */
 
-function Cart ({ cartItems, removeFromCartProp, totalCostProp, discountRateProp, bonusItems }) {
-// add removeFromCartProp fx parameter, and parameters that are variables that hold totalCost and discountRate. All this variable stuff will be in App function and used inside this function
-// Step 7: Added bonusItems to be used as prop on App.jsx, but before using it as a prop create conditional (wrap it as a function)
-// Step 7a: create callback fx to generate bonus items 
-const getBonusItemsToDisplay = () => {
-  // making it a switch statement will probably make code longer bc gotta create statements that will be true for each case to put inside switch ()
-  if (totalCostProp > 1000) {
-    return bonusItems;
-  } else if (totalCostProp >= 500) {
-    return bonusItems.slice(0, 3);
-  } else if (totalCostProp >= 300) {
-    return bonusItems.slice(0, 2);
-  } else if (totalCostProp >= 100) {
-    return [bonusItems[0]];
-  } else {
-    return [];
-  }
-};
-// Determine the bonus items to display (created this variabe so output can be used to iterate through in .map fx in the return (to append and update on page))
-const displayedBonusItems = getBonusItemsToDisplay();
-  return (
-    <section className="cart">
-      <h2>Cart</h2>
-      <h5>Discount: {discountRateProp}%</h5>
-      <h4>Total Cost: ${totalCostProp.toFixed(2)}</h4>
-      <ol>
-        {cartItems.map((item, index) => (
-          <li key={index}>
-            {item.name} - ${item.amount.toFixed(2)}
-            <button onClick={() => removeFromCartProp(item.id)}>Delete</button>
-          </li>
-        ))}
-      </ol>
-      {displayedBonusItems.length > 0 && (
-        <div>
-          <h3>Bonus Items</h3>
-          <ul className="bonuses">
-            {displayedBonusItems.map((bonus, index) => (
-              <li key={index}>{bonus}</li>
-            ))}
-          </ul>
-        </div>
-        )}
-    </section>
-  );
-}
+// function Cart ({ cartItems, removeFromCartProp, totalCostProp, discountRateProp, bonusItems }) {
+// // add removeFromCartProp fx parameter, and parameters that are variables that hold totalCost and discountRate. All this variable stuff will be in App function and used inside this function
+// // Step 7: Added bonusItems to be used as prop on App.jsx, but before using it as a prop create conditional (wrap it as a function)
+// // Step 7a: create callback fx to generate bonus items 
+// const getBonusItemsToDisplay = () => {
+//   // making it a switch statement will probably make code longer bc gotta create statements that will be true for each case to put inside switch ()
+//   if (totalCostProp > 1000) {
+//     return bonusItems;
+//   } else if (totalCostProp >= 500) {
+//     return bonusItems.slice(0, 3);
+//   } else if (totalCostProp >= 300) {
+//     return bonusItems.slice(0, 2);
+//   } else if (totalCostProp >= 100) {
+//     return [bonusItems[0]];
+//   } else {
+//     return [];
+//   }
+// };
+// // Determine the bonus items to display (created this variabe so output can be used to iterate through in .map fx in the return (to append and update on page))
+// const displayedBonusItems = getBonusItemsToDisplay();
+//   return (
+//     <section className="cart">
+//       <h2>Cart</h2>
+//       <h5>Discount: {discountRateProp}%</h5>
+//       <h4>Total Cost: ${totalCostProp.toFixed(2)}</h4>
+//       <ol>
+//         {cartItems.map((item, index) => (
+//           <li key={index}>
+//             {item.name} - ${item.amount.toFixed(2)}
+//             <button onClick={() => removeFromCartProp(item.id)}>Delete</button>
+//           </li>
+//         ))}
+//       </ol>
+//       {displayedBonusItems.length > 0 && (
+//         <div>
+//           <h3>Bonus Items</h3>
+//           <ul className="bonuses">
+//             {displayedBonusItems.map((bonus, index) => (
+//               <li key={index}>{bonus}</li>
+//             ))}
+//           </ul>
+//         </div>
+//         )}
+//     </section>
+//   );
+// }
 
 // Step 6: Create Checkout function handlesubmit fx must be created inside this one so that we can then put Checkout fx on <form onSubmit={Checkout}> element in jsx return. resetCart(); callback must be added right before return so form clears upon submission
 // Step 6a: Usestates for this function must be created so input gets tracked
@@ -115,25 +117,23 @@ return (
   <section className='checkout'>
     <h2>Check Out</h2>
       <form onSubmit={handleSubmit}>
-        <label htmlFor="firstName">First Name</label>
-        <input onChange={handleInputChange} value={formData.firstName} id="firstName" name="firstName" type="text" autoComplete="given-name"/>
-
-        <label htmlFor="lastName">Last Name</label>
-        <input onChange={handleInputChange} value={formData.lastName} id="lastName" name="lastName" type="text" autoComplete="family-name"/>
-
-        <label htmlFor="email">E-mail</label>
-        <input onChange={handleInputChange} value={formData.email} id="email" name="email" type="email" autoComplete="email"/>
-
-        <label htmlFor="zipCode">Zip Code</label>
-        <input onChange={handleInputChange} value={formData.zipCode} id="zipCode" name="zipCode" type="number" autoComplete="postal-code"/>
-
+        <label htmlFor="firstName"/>
+        First Name<br></br>
+        <input onChange={handleInputChange} value={formData.firstName} id="firstName" name="firstName" type="text" autoComplete="given-name"/><br></br>
+        <label htmlFor="lastName"/>
+        Last Name<br></br>
+        <input onChange={handleInputChange} value={formData.lastName} id="lastName" name="lastName" type="text" autoComplete="family-name"/><br></br>
+        <label htmlFor="email"/>
+        E-mail<br></br>
+        <input onChange={handleInputChange} value={formData.email} id="email" name="email" type="email" autoComplete="email"/><br></br>
+        <label htmlFor="zipCode"/>
+        Zip Code<br></br>
+        <input onChange={handleInputChange} value={formData.zipCode} id="zipCode" name="zipCode" type="number" autoComplete="postal-code"/><br></br>
         <button type='submit'>Submit</button>
       </form>
   </section>
   )
 }
-
-
 
 function App() {
   const [cartItems, setCartItems] = useState([]);
