+++ src/components/Cards.jsx
@@ -9,3 +9,19 @@
 // };
 
 // export default Cards;
+
+function Cards ({ birdData, addToCartProp }) { // name of prop is addToCart
+    return (
+      <section className='birds'>
+        {birdData.map((bird) => (
+          <div key={bird.id} className="card">
+            <h3>{bird.name}</h3>
+            <img src={bird.img} alt={bird.name} />
+            <p>${bird.amount.toFixed(2)}</p>
+            <button onClick={() => addToCartProp(bird)}>Adopt</button>
+          </div>
+        ))}
+      </section>
+    );
+  }
+export default Cards;
